Refact: Move decoder_tools to inside block, assuming the only function is to block decode
