+++ src/block/mod.rs
@@ -1,3 +1,4 @@
 pub mod random;
 pub mod breakers;
 pub mod generator;
+pub mod decoder;
