+++ src/lib.rs
@@ -1,4 +1,3 @@
 pub mod transaction;
 pub mod regtest_pack;
-pub mod decoder_tools;
 pub mod block;
